tests/requirements.txt
spacy-nightly>=3.0.0.dev13,<3.1.0
pytest<7.0.0
